requirements.txt
aiogram==3.1.1
telethon==1.34.0
aiosqlite==0.19.0
python-dotenv==1.0.0
aiofiles==23.1.0
aiohttp==3.9.1
